removed no longer used API method
