+++ web/courses.go
@@ -127,26 +127,6 @@ func (s *AutograderService) updateEnrollments(ctx context.Context, sc scm.SCM, c
 	return nil
 }
 
-func (s *AutograderService) getOrganization(ctx context.Context, sc scm.SCM, org string, user string) (*pb.Organization, error) {
-	gitOrg, err := sc.GetOrganization(ctx, &scm.GetOrgOptions{Name: org, Username: user})
-	if err != nil {
-		return nil, err
-	}
-	// check payment plan
-	if gitOrg.GetPaymentPlan() == FreeOrgPlan {
-		return nil, ErrFreePlan
-	}
-	// check course repos
-	repos, err := sc.GetRepositories(ctx, gitOrg)
-	if err != nil {
-		return nil, err
-	}
-	if isDirty(repos) {
-		return nil, ErrAlreadyExists
-	}
-	return gitOrg, nil
-}
-
 func updateReposAndTeams(ctx context.Context, sc scm.SCM, course *pb.Course, login string, state pb.Enrollment_UserStatus) (*scm.Repository, error) {
 	org, err := sc.GetOrganization(ctx, &scm.GetOrgOptions{ID: course.OrganizationID})
 	if err != nil {
